Move pv tests to own file (#352)
